One already previously addressed issue revisited:
As mentioned before, both line/surface Neumann conditions in 2/3-D and surface/volume Neumann conditions (i.e., body forces effectively) had been divided by density in their calculation until recently. For the low-Mach-number solver, from now on, surface/volume Neumann conditions will be multiplied by density according to the usual right-hand-side form rho*f, where f may, for example, be a gravity force. However, as I was able to learn, for some applications, more "structure-oriented" forces rather than "fluid-oriented" volume forces are required. Christiane already changed the respective part for impedance boundary conditions yesterday. Impedance boundary conditions are currently not supposed to be an element of low-Mach-number applications.

For "general" Neumann conditions, however, the "fluid-oriented" definition is kept alive also for the incompressible case, for the time being, that is, a gravity force, for example, is indeed a gravity force in the input file and does not have to be artificially multiplied by density before inserting the respective value into the input file, merely to be subsequently divided by density again during the simulation. This should be taken into account when using surface/volume Neumann conditions as body forces in 2/3-D in the incompressible case. Of course, as also mentioned before, it does not matter for (usual) incompressible flows with unity-value density.

However, if there is a need or desire for changing Neumann conditions in the incompressible case to the more "structure-oriented" definition, please let me know. In the present commit, the respective "infrastructure" for this purpose is already initialized for Neumann boundary conditions, though commented out, for the time being. It may easily be activitated, and transferred to body forces as well. The definition of "vede"-vectors relevant for the low-Mach-number case, on the one hand, and a constant scalar density_ relevant for the incompressible case, on the other hand, may allow for "simultaneously" multiplying and dividing by density and being able to obtain the correct value, though sounding odd in the first instance.

Further additions:

- added outflow stabilization for 2D fluid

- added potential spatial functional dependency of Neumann boundary conditions for 3D fluid such that, for instance, atmospheric stress conditions may be applied

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@7834 060e22d8-9d1c-0410-868a-e3a950d63d95