Corrected check for positive Element-Jacobians at wedge-nodes. Implemented stricter positive Jacobian test at gps for Dat-file validation.
I think we should push as much validation as possible to pre_exodus instead of discovering bad elements during baci-run. Keep in mind that many solid-elements do not even test for negative Jacos any more.
However, this might be worth a discussion - thinks a guy who is juggling with bio-geos....

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@7838 060e22d8-9d1c-0410-868a-e3a950d63d95