Epetra_FECrsMatrix introduced in LINALG::SparseMatrix:
So far, the implementation of the LINALG::SparseMatrix was only based on
an Epetra_CrsMatrix, which was able to handle (assemble) values, 
which are owned by the calling processor.
In addition to the Epetra_CrsMatrix, the LINALG::Sparsematrix can now work
on an underlying Epetra_FECrsMatrix.
The Epetra_FECrsMatrix delivers the necessary functionality
to assemble NON-LOCAL values as well.


For users of LINALG::SparseMatrix based on an Epetra_CrsMatrix
for assembling local values:

->	There are NO changes to be considered !
	All SparseMatrix constructors create Epetra_CrsMatrices by default.


For users of LINALG::SparseMatrix based on an Epetra_FECrsMatrix
for assembling NON-local values:

->	The matrix type has to be changed to FE_Matrix in a
	LINALG::SparseMatrix constructor in order to create
	an underlying Epetra_FECrsMatrix.
	Instead of Assemble(), FEAssemble() has to be called, 
	in all Evaluate()-methods, which fill the LINALG::SparseMatrix.
	Calling FEAssemble() causes internally the GlobalAssemble()-method to 
	distribute the nonlocal values to the owning 
	processors, before FillComplete() is called on the matrix. 
	Since Epetra_FECrsMatrix is a sub-class of Epetra_CrsMatrix,
	all Epetra_CrsMatrix functionality is also available.
	

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@7896 060e22d8-9d1c-0410-868a-e3a950d63d95