implemented stress/DLM condensation in intersected elements (it's exactly the same approach as in the solid hex8 EAS element   - thanks to Moritz for explaining the implementation.)
result:
- speedup + less memory consumption due to less global unknowns
- big speedup due to the use of MLFLUID2 (thanks Michael for explaining again all the gory details of AMG and its parameters)

In fact, condensing the stress unknowns works so well that - with this commit - it becomes the standard way of doing weak Dirichlet conditions in XFEM, although one can turn on the old way of solving nodal (u,p) and stress unknowns together by setting the parameter "DLM_condensation" to false. Right now, "DLM_condensation" is hardcoded to true.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@7963 060e22d8-9d1c-0410-868a-e3a950d63d95