Changed the NeoHooke, the AAA and the StVenent (3D version) material from Epetra_SerialDense to FixedSizeMatrix<T> objects
Created an intermediate layer of wrappers FixedSize->Epetra->FixedSize that allows smooth transition of all materials to FixedSizeMatrices

Once all materials have been transformed to FixedSizeMatrix<T> the intermediate layer can go away and we will be pure FixedSizeMatrix in materials. (have to talk to Lena first, though)

Eliminated the use of the remaining EpetraSerialDense matrices from the element sotet4, so_weg6 and so_hex8 (except for eas) that now use the lightweight wrappers in the material interface

Uniform nodal tet is to come next week.

Though using this twice lightweight wrapping of objects I gained a speedup of 10% per time step using the AAA and NeoHooke. This means the element speedup is even higher as usually the solver takes 50% of the time step size. Also, AAA and NeoHooke are really simple materials where there is not a lot of computation. Speedups for complex materials will be even larger.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@8018 060e22d8-9d1c-0410-868a-e3a950d63d95