Added QHULL_INC and QHULL_LIB to cluster config files.
As on the PCs (FC6 and FC8), qhull is searched for in the standard PATH (hence QHULL_LIB is empty). To use the qhull library, one has to set the QHULL flag in the defines file as usual.

qhull then be downloaded from the website and installed in the users account on the cluster

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@8078 060e22d8-9d1c-0410-868a-e3a950d63d95