Implemented another version of multi point constraint in 3D: When applied to a DSURF the displacements in a specified direction will be enforced to be equal for all nodes of this entity.
Input line:
-----------------------------DESIGN SURFACE NORMALDIR MULTIPNT CONSTRAINT 3D
DSURF  1
// ConditionID inittime masternode direction_x direction_y direction_z
E 4 - 1 0.0 1 1.0 0.0 0.0

"masternode" is defined relative to the nodal cloud.

Nightly test will follow.
--

Remark1:
As it is true for all MPC applied to a large number of nodes, this boundary condition should be combined with "UZAWAALGO  direct". This requires the usage of a direct solver as Superlu or UMFPACK restricting the models to a moderate number of DOFs.

Remark2:
The approach taken here is not the most efficient one could think of. The couling of the dofs is done by introducing algebraic constraints. It was only done because the whole framework was already there and only minor changes had to be made.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@8160 060e22d8-9d1c-0410-868a-e3a950d63d95