memory-efficient version of
void DRT::Discretization::BuildVolumesinCondition(
     const string name,
     RefCountPtr<DRT::Condition> cond)

the initial version cloned all volumes attached to every node

in cases where the condition is on every node, we required space
for up to 8 times the number of elements; space for 7 elements
was freed in the code, but the running process kept the memory
(that means I was not able to see the freed memory monitoring my 
ressources in top)

The current version kept the design of the initial one. 
The major difference: We do not store cloned volumes in a multimap 
anymore, just a volume identifier consisting of the gid of the
element which generated the voume and the number of the volume 
generated by this element

after erasing multiple identifiers from the list, we clone
the only the remaining volumes.







git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@8167 060e22d8-9d1c-0410-868a-e3a950d63d95