Computation of mean values of nodal/cp quantities. The means are computed as time averages.
An additional method to do an averaging in a homogeneous, axis 
aligned, direction is provided. A requirement for this averaging
procedure is that all nodes are repeated in identical planes
along the axis along we do the averaging (like it is the case
for a flow around a square cylinder, a backward facing step,
a diffuser or a plane channel).  
--- all space averaged information is distributed back to all 
nodes (for easy visualisation in paraview). Note that this 
involves heavy communication --- including two round robin 
loops --- and produces additional output.

Up yo now, there's no connection between this method and the 
remaining code. This will be done by the statistic_manager
(modifications will follow).




git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@8420 060e22d8-9d1c-0410-868a-e3a950d63d95