corrected calculation of angular momentum in 3D (lift&drag method):
computed distances refer now to an axis of rotation,
specified in the surface(!) lift'n'drag conditions.

Adapted two input files (which do not have a maintainer?!)
and activated calculation of lift'n'drag in these files.

For my rotating cylinder example, the relative error 
of computed angular momentum to the exact value is now around 1%
:-)



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@8443 060e22d8-9d1c-0410-868a-e3a950d63d95