Christmas cleaning of intersection, search tree and integration cell (Part 1)
Refactoring of intersection method

-> tested for tetrahedal volume and triangular surface elements 
(@Florian: the intersection is now ready to work with triangulated level stets)
-> optimized number of integration cells for plane elements
-> improved and extended templation -> moved methods to intersection_templates.H
-> numerous little clean and speed ups


Refactoring of search tree

-> tested for tetrahedal volume and triangular surface elements


Integration cell:

-> provided a more efficient implementation of the integration cell class
and coordinate transformations

@Axel: It is of course up to you to copy the intersection method even though you know that
it is under refactorisation. Unfortunately it is now completely outdated and the interface methods changed. 
Since your copy doesn t compile anymore, I removed it. 
If you want to add it again, please do so in the drt_xfem folder.
As I already explained to you, an arbitrary linear triangulation of bilinear or biquadratic quad-elements
with two linear triangular elements might lead to highly inaccurate results. 
I integrated the feature you are trying to implement for plane elements (see above). 
But be aware that this will only speed-up the computation of plane and 
undeformed elements (= all of your nightly test cases) but not realistic
simulations. In addition, I provided you with an own implementation 
of the simple container class integration cell,
which avoids several identical on-the-fly computations of various nodal coordinate transformations. 
This results in a runtime reduction of 30-40% (!) for element evalutions, also for deformed elements.
So, it seems to me that there are still a lot of possibilities to improve efficiency in your XFEM-code, 
instead of hardcoding an optimization only delivering accurate results for nightly test cases in the intersection method.

-> integration of changes into the code (hope that all necessary files found their way into the commit :-) )

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@8581 060e22d8-9d1c-0410-868a-e3a950d63d95