An implicit predictor for STI
How to call:
-------------------------------------------------STRUCTURAL DYNAMIC
PREDICT              TangDis


What does it do:
It predicts a displacement field D_{n+1}^{<0>} which follows the tangent of the last converged state such that the Dirichlet BCs are satisfied.
  D_{n+1}^{<0>} = D_{n} + K_{n,eff}^{-1} . (- R)
with
  R = F_{int,n} - F_{ext}(t_{n+1}) + K_{n} . (D_{n+1}-D_{n})|_{DBC}

The implementation _could_ be simplified by
  R = K_{n} . (D_{n+1}-D_{n})|_{DBC}
which appears funny, but remember K_{n,eff}^{-1} is not the inverse of K_{n} due to the application of the Dirichlet BCs (i.e. the reduction to the test space).


Where to use:
It is likely beneficial in cases of displacement-controlled, axial stretching of incompressible materials. The volume change due to prediction with "ConstDis" or "ConstDisVelAcc" is not focused solely on the elements attached to the Dirichlet boundary, but 'smeared' across body.


Where to ignore:
Systems in which DBCs are not prescribed, i.e. they are only constant. Or unsupported structures.


How to count iterations:
"TangDis" is an implicit predictor. Thus it calls the solver once. A fair comparison takes this into account and increases the iteration by 1.




git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@8642 060e22d8-9d1c-0410-868a-e3a950d63d95