changed f3_calc_means, i.e. the calculation of integral in plane averages for turbulent channel flow
Up to now it was required that every element in a layer 
was of the same size. Now we compute the Jacobian of each 
2d element which is generated by a wall parallel cut 
through the 3d element, that means we include different 
weights for elements of different sizes (necessary for 
deformed meshes).

The last remaining assumption is that all planes stay 
parallel to the wall (this property of the mesh motion 
is exploited when computing the jacobian of the area 
element)





git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@8711 060e22d8-9d1c-0410-868a-e3a950d63d95