For the Assemble method (CRS_Matrix), check, whether row dofs exist on this proc. If not, don't assemble!
Background: If no row dofs exist (happens in XFEM simulations - one proc might have just a hole/void),
the CRS Matrix does not ghost the coldofs as expected. In particular, the coldofmap of the discretization and the
Epetra CRSmatrix can be different in this case! This is ok, as long as we don't try to assemble in such a case.
Before, the DEBUG baci incorrectly fails, while the FAST baci
did ignore a test for coldofs and worked correctly.

Fix approved by M. Gee.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@8932 060e22d8-9d1c-0410-868a-e3a950d63d95