Added springs variant with name "springs_const_stiff"
It does not use the current configuration to compute the stiffness matrix, which makes the stiffness
matrix independent of the displacement.

While this might not allow the huge deformations possible with the standard springs, it does not create "bands" or similar things, when computations with oscillation structures are performed. So far, "springs_const_stiff" is the only mesh deformation algorithm that survives more than 5-10 cycles of the FSI3 benchmark (Currently, it is at 60 cycles and still running). When compared to classic_lin or laplace, it survives the quite strong mesh deformations at the beam tip without applying constraints or different materials to the "near-beam" fluid mesh (-> ease of use ;-) ).

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@8935 060e22d8-9d1c-0410-868a-e3a950d63d95