Restart for impedance BCs:
To be able to restart smoothly traction vector has to be stored in restart files

Also, when time step size is changed within restart, the history of flowrates has to be interpolated to the new Delta t intervals. when time step size is unchanged, nothing happens.

Upon change of time step size the number of impedances in frequency domain changes. Experimenting with time step sizes yields that flow/pressure curves are time step size dependent and can thereofre not match exactly in a restart with time setp size change.

Interpolating flow rates to the new time step size therefore is an approximation - a slight 'bump' in pressure has to be expected at every change of time step size. The bump though is now small enough for an FSI simulation to go through it.

Time step size can be made smaller or larger, time step size can be adjusted within first or subsequent periods.

It is recommended to choose the new dt to be an exact fraction or multiple of the old dt.

Now, FSI with Windkessel BCs becomes practically possible....

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@8979 060e22d8-9d1c-0410-868a-e3a950d63d95