added another "springs" mesh moving algorithm for ALE
The stiffness matrix is constant over time (see last committed springs variant) AND we compute the displacement relative to the reference configuration. This will probably limit the maximum mesh deformation when compared to the original springs algorithms.
On the positive side, mesh deformation can not diverge over time for periodic deformations.

The implementation is a copy of the "classic_lin" algorithm, but with a new element evaluate call for the springs stiffness computation.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@8982 060e22d8-9d1c-0410-868a-e3a950d63d95