When using ILU as smoother within ML multigrid, ML always did 2 ILU-preconditioned Richardson iterations (sweeps) due to a conflict between default parameters and our parameters that is now resolved.
As doing ILU sweeps is expensive and the benefit/cost ratio usually is not so good I fixed this to being 1 sweep of ILU now.

Users using ML with ILU (e.g. common in fluids) will experience an increase in aztec iterations with a decrease of time per iteration. In my tests overall solution time dropped 20-30% !

Newton behavior is not affected.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@9107 060e22d8-9d1c-0410-868a-e3a950d63d95