postprocessing for zero-sized elements in knotspan --------------------------------------------------
we want to keep them since we rely on the cartesian element id 
access functions from the knotvectors. If we would just skip them
(as it is done during element calls), we would loose some 
connectivity in the result displayed by paraview

As a work-around, we replace the zero-sized element with the next 
nonzero element --- this preserves the connectivity, and everything 
looks nice and smooth again.

Note: This work-around will not work in parallel (or a
      special ghosting for elements along interpolated
      boundaries has to be applied)

Note: The next nonzero element will be plotted twice



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@9249 060e22d8-9d1c-0410-868a-e3a950d63d95