ACE progress
* a small step for the code, a giant step for a greener and better ......
The FE simulation is dynamic, the MD simulation is dynamic. One would be very stupid not to try to
make the coupling dynamic as well. And here it goes. A first idea. Additionally to the deformation gradient the second derivative of the deformation gradient in the first characteristic direction is considered as well. Not yet connected down to MD (water examples to look into the dynamic coupling matter are on the way), but a first nice step. Forces still look a little bit high, but are in a reasonable range. So, let the water bouncing begin....



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@9334 060e22d8-9d1c-0410-868a-e3a950d63d95