added two functions as an optional argument for local coordinate systems.
1) Function one has to be vector-valued and prescribes a field of rotation axes 
   (for example the constant field (0,0,1) for a rotation around the z-axis)
2) Function two is scalar valued and provides the angle of rotation as a function
   in space.

The local coordinate system in one node is rotate around the axis defined by field
one by the angle taken from field two. This allows for example the definition of
coordinate systems around circular geometries.

Warning:
o This will not work for NURBS problems, since shapefunctions are never equal to 
  one and thus the rotation of one control point does not necessarily lead to a
  rotated 'dof' in the spatial sense
o There are problems with full rotations of coordinate systems in combination with
  the fast version by Burkhard and AZTEC preconditioners. If you use the fast version
  make sure that the trafo matrix has always nonzero diagonal elements (you can always
  provide such a trafo by chosing the right local coordinate system and rotate it 
  by less than 45 degrees)







git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@9384 060e22d8-9d1c-0410-868a-e3a950d63d95