increased warning level using -Wextra
Note: sorry for bothering you with a series of commits in the last 2 days - they all were for the purpose to reduce warning levels while searching for the xbug.
Since we now have almost warning free code at the Wextra level, I think it is good to keep this level, hence this commit.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@9441 060e22d8-9d1c-0410-868a-e3a950d63d95