Bugfix when reusing ML preconditioners in xfem runs with moving interfaces:
Somehow, the fluid solver caries over preconditioners or similar stuff to the new timestep, even though I told him to recompute everything in the first newton step of the fluid Newton loop  :-( ???

This crashes the code in the middle of some ML code in Trilinos. I guess, changing number of dofs is not handles properly...

Solution: Dump the solver_ and create a new solver_ at the beginning of each fluid iteration. Then ML with reuse of preconditioners works as expected meaning very well.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@9457 060e22d8-9d1c-0410-868a-e3a950d63d95