Isochoric splitting for sosh8p8 (with quadratic convergence):
Optional input flag at element line:
   [ ISO ( Mat | Enf ) ]

Mat : Material stress response and elasticity matrix is taken as delivered; default.
Enf : Isochoric content is ripped of material stress response and elasticity matrix.

Subversive CommiTheme (TM):
We are all agreed that your theory is crazy. The question which divides us is whether it is crazy enough to have a chance of being correct. My own feeling is that it is not crazy enough. (Niels Bohr)


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@9478 060e22d8-9d1c-0410-868a-e3a950d63d95