modified setup/restart process for monolithic fsi approaches.
---

Moved setup of coupling and creation of combined dofmap from the constructor
calls to an extra method called SetupSystem.

This is done in order to be able to read the restart before the combined maps 
are constructed. 

This is necessary since upon restart it is possible that a call to ReadMesh() 
calls FillComplete() via Redistribute() on one of the discretisations, resulting 
in new dofmaps and rendering the old maps invalid.




git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@9634 060e22d8-9d1c-0410-868a-e3a950d63d95