Added 3D diffusion element.
This is no production element! In particular it is not meant to be a substitute to SCATRA and friends! I just need an simple xfem element to play with element stiffness matrix and produce simple 3D xfem stiffness matrices for multigrid research. It might also be a sand box element to examine how to program other physics with xfem in baci.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@9671 060e22d8-9d1c-0410-868a-e3a950d63d95