An ANS method for sosh8p8. You use it by setting    ANS OnSpot on the respective element input line.
Application:
'ANS OnSpot' is a modifaction of the 'ordinary' ANS Later(al) method to prevent shear and trapezoidal locking. It's application is particular for meshes with one elemant across shell thickness. For these meshes, a more accurate solution is found cf 'ANS Later'. Its application is less recommended with more than 4 elements across thickness. Here, the (mild?) inconsistency dominates in and the solution is not going to converge to the correct result (but stays somewhat close). Use 'ANS Later' (or 'ANS None') instead which should be pretty close.

Background:
'ANS OnSpot' utilises additional terms in the ANS modified natural strains E23 and E13. The strains allow to express E23 and E13 bi-linearly (in r and s) in the shell centroidal plane. They a scaled proportional to the element aspect ratio. Essentially, the polynomial content of energy expressions of 'ANS None' are recovered, but, with swapped element dimenions such that shear locking does not occur with increased slenderness in thickness direction. Funny.

Warning:
In case you discover sosh8p8 flag 'EAS sosh8' and 'EAS atype': Resist. This is meant for experiments only.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@9739 060e22d8-9d1c-0410-868a-e3a950d63d95