Added analytic solution for 2D Jeffery-Hamel flow.
It is a 2D stationary flow problem including convection. The solution is defined as an differential equation in polar coordinates of the (scalar) radial velocity u(alpha) as

u'' + 4u + u^2 = 6C

The parameter C somehow governs the resulting Reynoldsnumber. One particular solution was computed and developed using Taylor-series and is coded as Function object in drt_function.H/.cpp. This means, it can be used as Dirichlet boundary condition and field function to compute, for example, the L_2 norm like sqrt(\int_\Omega |u^h - u^exact|^2 ).

Documentation will be extended shortly.

Thanks to Thomas K. for helping me with Mathematica.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@9845 060e22d8-9d1c-0410-868a-e3a950d63d95