strugenalpha now has a parameter "LOADLIN" which turns external load linearization on/off. Previously, external forces were computed in the predictor only and held constant during Newton. Now, external forces can be updated during Newton (at some price) such that the converged solution is actually w.r.t. true orthopressure load.
LOADLIN is currently implemented only in ConstantPredictor and FullNewton but I'll doo the other guys later.

LOADLIN is implemented in a TR fashion but there is also a mid-point implementation that is commented out (it did not converge as good as the TR, maybe there is still something wrong with it?). For the ConstantPredictor TR and mid-point of course are identical.

In general, I do not recommend using LOADLIN as the convergence benefit is small compared to the price in time. It's just that we now can say "We've tried load linearization. It does not matter"

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@9945 060e22d8-9d1c-0410-868a-e3a950d63d95