Added check for NaN's in the expression time curve, since while computing derivatives of CURVES, the automatic differentiation (SACADO package) fails for polynomial expressions as
CURVE1 on EXPR FUNC t^2 t1 0.0 t2 1.0

when evaluated at t=0. This is unfortunate because we wan't to start usually from t=0.0. However, I could not find the source of this problem. Note that

CURVE1 on EXPR FUNC 2^t t1 0.0 t2 1.0

works properly.

The work-around - after 2 days of hair pulling - is of course:

CURVE1 on EXPR FUNC t^2 t1 0.0000000000000000000000001 t2 1.0



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@9955 060e22d8-9d1c-0410-868a-e3a950d63d95