add line definitions that are used to read both time curves and spacial functions
These line definitions are not specific to time curves or functions. Thus the actual
reading code did not shrink too much. On the other hand, there is just one line
definition subsystem. It is meant to be used for element lines are well. Any maybe
the condition and material reading subsystems can be based on these line definitions.

In any case, the -d option now gives time curve and spacial function information
just as it is meant to be.

This is the first version. No comments yet. Expect bugs.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@9984 060e22d8-9d1c-0410-868a-e3a950d63d95