Switch fluid3 element to new input system. All other elements remain unchanged.
With this change the fluid3 input line is fixed. There is now a required format.
Nothing else will be accepted. The input format is printed if baci is called with
the -d switch.

Note: The number of gaussian points for fluid3 elements is not read from the dat file
anymore. Instead that number of gaussian points that matches the element shape is
chosen.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@10017 060e22d8-9d1c-0410-868a-e3a950d63d95