Some cleanup on test files again.
The point here is to remove unused fluid and ale element flags. These are automatic
modifications that touch all matching files regardless of there content. If any of
the old ccrarat test cases failes due to these changes, it will be removed. We do not
maintain the old C code any longer.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@10025 060e22d8-9d1c-0410-868a-e3a950d63d95