Add structure wrapper and handle NOX correction in additional wrapper class.
Note: There is a subtil semantic change in ADAPTER::Structure::Evaluate(disp).
Up to now disp was supposed to be d^{n+1}_{i+1} as that is what NOX calculates.
Now Evaluate expects the true Newton increment. The translation is managed by
StructureNOXCorrectionWrapper.

The reason for that change is Axels monolithic XFSI implementations, that does
not use NOX.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@10032 060e22d8-9d1c-0410-868a-e3a950d63d95