new functions to handle Dirichlet boundaries with XFEM
EvaluateDirichletXFEM()
DoDirichletConditionXFEM()

These functions are modified copies of the corresponding standard functions.

The problem of XFEM Dirichlet boundaries for XFEM has not been solved in BACI. We are not able to handle Dirichlet boundary conditions if the number of degrees of freedom on nodes varies along the boundary. Generally speaking, discontinuities treated with the XFEM approach must not touch the boundary, so far.
This commit does not close this gap. It is merely a hack to circumvent restrictions on the number of dofs on boundaries. We rely in the numbering of standard and enriched degrees of freedom. XFEM Dirichlet boundary conditions remain an open issue in BACI.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@10157 060e22d8-9d1c-0410-868a-e3a950d63d95