Implemented basic functionality for 3D second-order mortar contact using a penalty method / augmented Lagrangian method (frictionless)
- second commit
- we now use piecewise linear interpolation for constraint discretization
  (Lagrange multiplier test functions) but still the original quadratic
  interpolation for contact virtual work (Lagrange multiplier interpolation functions)
- this can be interpreted as a kind of Petrov-Galerkin approach
- results look very good now
  (+) multi-element contact patch tests for hex20, tet10 are now passed
  (+) no strange zig-zag deformations in the contact interface anymore

@Markus, Michael:
This version now should be OK to demonstrate our general capability of doing 3D second-order contact to RollsRoyce. Friction needs to be tackled next. Of course, it will still be interesting to investigate two things:

(a) Why doesn't my fully piecewise linear approach (as suggested by Puso and Laursen) work properly? Maybe, it's just a bug, but maybe it's also a problem with the whole idea...

(b) How can the 3D second-order case with DUAL Lagrange multipliers be treated? Of course, this is what we
actually want in the end. I am working on this with my HiWi at the moment...


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@10292 060e22d8-9d1c-0410-868a-e3a950d63d95