Enabled dynamic load balancing in ML using parmetis.
This is an extremely important performance enhancement for the AMG preconditioned iterative solvers. It allows coarse grids to be redistributed to a subset of processors once a certain minimum size threshold of rows per processor is no longer met. In practice, this means that e.g. fine grid is run on 12 procs, medium grid on 9 procs and coarsest grid on 1 proc. 

A lot of communication overhead is therefore saved on coarser grids.

The feature is automatically switched on with the define PARMETIS in baci since parmetis is required by ML to do the load balancing.
There are no parameters (yes, really !!) to choose that influence load balancing behavior.

In case you run large examples on lots of processors using AMG-Aztec as a solver this is a must have new feature, so put PARMETIS in your defines file.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@10379 060e22d8-9d1c-0410-868a-e3a950d63d95