Included FSI version with sliding ALE-Structure interface for large rotations of the structure field in FSI. This commit is based on the Georg Hammerl's student project, but was extended to 3d simulations and parallel computations.
Feature can be activated by the following dat-file input in FSI DYNAMIC
PARTITIONED                     DirichletNeumannSlipALE
SLIDEALEPROJ                    Ref/Curr

The parameter SLIDEALEPROJ determines the starting point for the projection of the ALE interface nodes on the structure interface. For 'Ref' the reference configuration + average interface translation is used, for 'Curr' the last converged configuration + average interface translation increment.


Work is still rather experimental and has to be discussed with Ulrich.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@10669 060e22d8-9d1c-0410-868a-e3a950d63d95