Drastically(!) removed memory consumption needed for computing a matrix-matrix-product C = A*B of two global matrices A and B in BACI. Thanks to Michael for the discussion this morning.
For the result matrix C, a (worst case) guess for the bandwith was used up to now:
  expected number of non-zeros per matrix row of C = bandwidth(A)*bandwidth(B).
This lead to a massive allocation of memory for expected non-zero entries.
(In my example: bandwidth(A)=153,bandwidth(B)=51 => allocation of 7803 doubles per matrix row of C )

As a consequence, multiplication routines use now a better initial guess for the bandwith.
For my specific example:
Before: 70.7% memory usage on my machine
After:  42.0% memory usage (LINALG::Multiply)
        24.2% memory usage (LINALG::MLMultiply)

Some remarks:
1) This affects all users of LINALG::Multiply and LINALG::MLMultiply
   (By the way, there are 4-5 different multiplication methods around at the moment)
   
2) Our SparseMatrix class is currently NOT calling the OptimizeStorage() method of the EpetraCrs_Matrix.
   Thus, memory that was allocated too much during creation of an instance of SparseMatrix is not freed
   during the Complete() call on the matrix, but kept as it is. This is another bug that has to
   be fixed by some experts.-> put that topic into our bugtracing system Mantis
    


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@10680 060e22d8-9d1c-0410-868a-e3a950d63d95