This commit contains: 1 generalized one of the multipoint constraints 2 additional versions for monolithic fsi calculation with algebraically constrained structure fields 3 changed constraint world from SparseMatrix to SparseOperator
to 1:
SURFACE NORMALDIR MULTIPNT CONSTRAINT 3D can now be used to constrain the relative displacement with respect to a given direction or the relative current configuration with respect to this direction. Furthermore condition now accepts timecurves and values can be constrained relative to an initial value.
Adapted nightly test.

to 2:
Implemented (with the help of Lena) something similar to Lena's lung-fsi algorithm to be used for fsi with constrained structure fields. Instead of merging structure and constraint matrices together within a special structure adapter, the work is now done in a global special BlockSparseMatrix. Preconditioning is done with a Schur complement.

New input flags:
COUPALGO                        iter_constr_monolithicstructuresplit
COUPALGO                        iter_constr_monolithicfluidsplit

to 3:
The constraint manager can now deal with SparseOperators and does not rely on SparseMatrix any more. 

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@10940 060e22d8-9d1c-0410-868a-e3a950d63d95