Bunch of changes related to lung fsi simulations ================================================
Structure
---------
There are no forces on the outflow boundary (see previous commit). However,
in order to obtain a reasonable orientation of forces at the fsi part of the outflow
boundary, also the adjacent elements on this boundary need to be included. 
No distinction between "inflow" and "other" parts of the boundary is necessary
anymore (since the undesired contributions are just zeroed out based on a map
extractor introduced for this purpose) -> removed this distinction.

Fluid
-----
There is no contribution of the fsi part of the outflow boundary to the overall
flow rate (u - u_grid = 0) -> zeroed out.

Lung FSI algorithm
------------------
Included constraint part in first evaluation of rhs in lung fsi simulations.
Furthermore, the parameter theta used for time integration of flow rates was
chosen equally to the one used in general fluid time integration (however, currently
hard coded).

Output
------
Enabled output of constraint related forces in lung fsi simulations. For this purpose,
the DiscretizationWriter needs to be accessible in all fluid adapters and underlying
time integrators.



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@10965 060e22d8-9d1c-0410-868a-e3a950d63d95